Move hour tick marks to Bezel
